code refactoring: create TokenProvider
